backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.7.2
googlemaps==4.10.0
gspread==6.1.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
Pillow==10.4.0
gunicorn==22.0.0
